requirements.txt
streamlit
PyMuPDF
python-docx
spacy
faiss-cpu
sentence-transformers
transformers
torch
